+++ front/src/dashboard/slips/RouteSlipsView.module.scss
